finally backend over

+++ 11.py
@@ -1,39 +1,34 @@
 import json
-store={
-     "apple":{"weight":100,"price":100},
-     "banana":{"weight":100,"price":100},
-     "rice":{"weight":100,"price":100},
-     "milk":{"weight":100,"price":100}   
+import os
+store = {
+    "apple": {"weight": 100, "price": 100},
+    "banana": {"weight": 100, "price": 100},
+    "rice": {"weight": 100, "price": 100},
+    "milk": {"weight": 100, "price": 100}
 }
-print(store)
 def store_input():
-    with open("store.json","w") as f:
-        json.dump(store,f)
+    with open("store.json", "w") as f:
+        json.dump(store, f)
 def load_store():
-    try:
+    if not os.path.exists("store.json"):
+        with open("store.json", "w") as f:
+            json.dump(store, f)
+        return store
+    else:
         with open("store.json", "r") as f:
             return json.load(f)
-    except:
-        return {
-            "apple":{"weight":100,"price":700},
-            "banana":{"weight":100,"price":700},
-            "rice":{"weight":100,"price":700},
-            "milk":{"weight":100,"price":700}  
-        }
-store=load_store()
-bill={}
+store = load_store()
+bill = {}
 def shopping():
     while True:
-        print("1.Add items\n 2. View Bill\n 3. Exit")
-        choice=int(input("Enter your choice:"))
-        if choice==1:
-            item=input("enter item name:").lower()
+        print("\n1.Add items\n2.View Bill\n3.Exit")
+        choice = int(input("Enter your choice:"))
+        if choice == 1:
+            item = input("Enter item name:").lower()
             if item in store:
-                weight = float(input("enter weight:"))
-
+                weight = float(input("Enter weight:"))
                 if weight <= store[item]["weight"]:
                     total_price = weight * store[item]["price"]
-
                     store[item]["weight"] -= weight
                     store_input()
                     if item in bill:
@@ -44,23 +39,20 @@ def shopping():
                             "weight": weight,
                             "price": total_price
                         }
-
-                        print(f"{item} added. weight={weight} Price = {total_price}")
-
+                    print(f"{item} added: {weight}kg: ₹{total_price}")
+                else:
+                    print("Not enough stock")
             else:
-                print("Not enough stock!")
-
+                print("Item not found!")
         elif choice == 2:
-             print("\n--- BILL ---")
-             total_amount = 0
-             for item, details in bill.items():
+            print("\n--bill--")
+            total_amount = 0
+            for item, details in bill.items():
                 print(f"{item} - {details['weight']}kg - ₹{details['price']}")
                 total_amount += details['price']
-             print("Total =", total_amount)
-        
-        else:
-            print("Thank you for shopping with us")
+            print("Total =", total_amount)
+        elif choice == 3:
+            print("Thank you for shopping")
             break
+
 shopping()
-
-
